MD3  - shaders are now processed  - multi-part player models are handled correctly
Material system
 - added flags for 'usealpha' or 'ignorealpha' setting of textures

LWO
 - fixed texture assignment bug due to invalid tag list

3DS
 - improved handling of dummy nodes

Viewer:
 - lines&points are now displayed
 - improved animation control via slider (still some stuff missing)
 - skeleton is now displayed
 - some other minor fixes

Validator:
 - some material issues are warnings now (no errors anymore)

git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@349 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/LWOLoader.h
@@ -455,7 +455,7 @@ inline void LWOImporter::GetS0(std::string& out,unsigned int max)
 		}
 		++mFileBuffer;
 	}
-	unsigned int len = (unsigned int) ((const char*)mFileBuffer-sz);
+	size_t len = (size_t) ((const char*)mFileBuffer-sz);
 	out = std::string(sz,len);
 	mFileBuffer += (len&0x1 ? 1 : 2);
 }
